Got tearing working
The previous bug that was causing the particle to not get process was because I never set the constraint connection count of each particle.

Right now the tearing is based on the distance from the original restitution distance, but right now it's using a constant multiplier for all constraints, meaning that the constraints that are further apart (in terms of which particles they are connected to) are more resistance to tearing.