Got cloth tearing to behave better
I turned down the mass of the particles.

I made it so each constraint has a random tearing resistance value between 0.1 and 2.1 cause why not.

Pinned particles have their constraints tearing resistance to always be 3.0.

Might have fixed a bug with some constraints still working after a particle is suppose to be dead. (Not sure)

The 'h' button can now be used to light the cloth on fire.
Although I need to use a better random generator when selecting which particle to set fire to.
I also need to make INPUT_DOWN_FIRST work for the 'h' key.